# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COMMAND Social Media Scheduling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Post {
  id: string;
  brand: string;
  platforms: string[];
  caption: string;
  image_url?: string | null;
  video_url?: string | null;
  media_format: string;
  link_url?: string | null;
  link_url_tagged?: string | null;
  content_type: string;
  headline_variant?: string | null;
  scheduled_at: string;
  status: string;
  posted_at?: string | null;
  error_message?: string | null;
  fb_post_id?: string | null;
  ig_post_id?: string | null;
  gbp_post_id?: string | null;
  created_at: string;
}

export interface CreatePostRequest {
  brand: string;
  platforms: string[];
  caption: string;
  image_url?: string | null;
  video_url?: string | null;
  media_format?: string;
  link_url?: string | null;
  content_type: string;
  headline_variant?: string | null;
  scheduled_at: string;
  post_now?: boolean;
}

export interface PostMetric {
  id: string;
  post_id: string;
  platform: string;
  fetched_at: string;
  likes: number;
  comments: number;
  shares: number;
  reach: number;
  impressions: number;
  clicks: number;
  engagement_rate: number;
  created_at: string;
}

export interface RankedPost {
  id: string;
  brand: string;
  platform: string;
  content_type: string;
  headline_variant?: string | null;
  reach: number;
  likes: number;
  comments: number;
  engagement_rate: number;
  posted_at?: string | null;
}

export interface MetricsSummary {
  totalPosts: number;
  avgEngagementRate: number;
  bestPlatform?: string | null;
  bestContentType?: string | null;
  rankedPosts: RankedPost[];
}

export interface HeatmapCell {
  day: number;
  hour: number;
  avgEngagement: number;
  postCount: number;
}

export interface Asset {
  id: string;
  brand: string;
  url: string;
  format: string;
  content_type: string;
  times_used: number;
  avg_engagement?: number | null;
  created_at: string;
}

export interface Brand {
  id: string;
  brand: string;
  display_name: string;
  emoji: string;
  fb_page_id?: string | null;
  ig_account_id?: string | null;
  gbp_location_id?: string | null;
  brand_voice: string;
  created_at: string;
}

export type ListPostsParams = {
  brand?: string;
  status?: string;
  platform?: string;
  content_type?: string;
  limit?: number;
  offset?: number;
};

export type ListPosts200 = {
  posts: Post[];
  total: number;
};

export type GetMetricsSummaryParams = {
  brand?: string;
  platform?: string;
  content_type?: string;
  date_from?: string;
  date_to?: string;
};

export type GetInsightsParams = {
  brand?: string;
};

export type GetInsights200 = {
  insights: string[];
};

export type GetHeatmapParams = {
  brand?: string;
};

export type GetHeatmap200 = {
  data: HeatmapCell[];
  hasEnoughData: boolean;
};

export type ListAssetsParams = {
  brand?: string;
  content_type?: string;
  format?: string;
};

export type ListAssets200 = {
  assets: Asset[];
};

export type GetAssetUploadUrlBody = {
  filename: string;
  contentType: string;
  brand: string;
  assetContentType: string;
  format: string;
};

export type GetAssetUploadUrl200 = {
  uploadUrl: string;
  publicUrl: string;
  assetId: string;
};

export type ListBrands200 = {
  brands: Brand[];
};

export type GenerateCaptionBody = {
  brand: string;
  content_type: string;
  brief: string;
};

export type GenerateCaption200 = {
  caption: string;
};

export type GetBestTimesParams = {
  brand?: string;
};

export type GetBestTimes200TimesItem = {
  day: string;
  hour: number;
  label: string;
  isGeneric: boolean;
};

export type GetBestTimes200 = {
  times: GetBestTimes200TimesItem[];
  hasEnoughData: boolean;
};