# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COMMAND Social Media Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: string;
  brand: string;
  url: string;
  format: string;
  content_type: string;
  times_used: number;
  avg_engagement?: number | null;
  created_at: string;
}