# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COMMAND Social Media Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface Brand {
  id: string;
  brand: string;
  display_name: string;
  emoji: string;
  fb_page_id?: string | null;
  ig_account_id?: string | null;
  gbp_location_id?: string | null;
  brand_voice: string;
  created_at: string;
}